clean code: move import into ops
