+++ ops/op_image_io.py
@@ -1,8 +1,7 @@
 import bpy
 import os
-
+import sys
 from bpy.props import StringProperty, BoolProperty, EnumProperty
-
 
 class SPIO_OT_import_image(bpy.types.Operator):
     """Import all image as reference (Empty object)"""
@@ -83,10 +82,83 @@ class SPIO_OT_import_image(bpy.types.Operator):
 
         return {'FINISHED'}
 
+if sys.platform == "win32":
+    from ..clipboard.windows import PowerShellClipboard as Clipboard
+elif sys.platform == "darwin":
+    from ..clipboard.darwin.mac import MacClipboard as Clipboard
+
+
+class ImageCopyDefault:
+    @classmethod
+    def poll(_cls, context):
+        if sys.platform in {"darwin", "win32"}:
+            return (
+                    context.area.type == "IMAGE_EDITOR"
+                    and context.area.spaces.active.image is not None
+                    and context.area.spaces.active.image.has_data is True
+            )
+
+    attr = [
+        'file_format',
+        'color_mode',
+        'color_depth',
+        'compression',
+    ]
+
+    def set_format(self, restore=False):
+        settings = bpy.context.scene.render.image_settings
+        for a in self.attr:
+            if restore:
+                setattr(settings, a, getattr(self, a))
+            else:
+                setattr(self, a, getattr(settings, a))
+
+    action = 'pixel'
+
+    def execute(self, context):
+        active_image = context.area.spaces.active.image
+        image_path = os.path.join(bpy.app.tempdir, f'{active_image.name}.png')
+
+        self.set_format()
+
+        bpy.ops.image.save_as(filepath=image_path, save_as_render=True, copy=True)
+        # push to clipboard
+        clipboard = Clipboard()
+        if self.action == 'pixel':
+            clipboard.push_pixel_to_clipboard(path=image_path)
+        else:
+            clipboard.push_to_clipboard(paths=[image_path])
+
+        self.set_format(restore=True)
+
+        self.report({'INFO'}, f'{active_image.name} has been copied to Clipboard')
+
+        return {'FINISHED'}
+
+
+class SPIO_OT_export_pixel(ImageCopyDefault, bpy.types.Operator):
+    """Export as Image Pixel"""
+    bl_idname = 'spio.export_pixel'
+    bl_label = 'Copy Pixel'
+
+    action = 'pixel'
+
+
+class SPIO_OT_export_image(ImageCopyDefault, bpy.types.Operator):
+    """Export as Image File"""
+    bl_idname = 'spio.export_image'
+    bl_label = 'Copy Image'
+
+    action = 'file'
+
 
 def register():
     bpy.utils.register_class(SPIO_OT_import_image)
+    bpy.utils.register_class(SPIO_OT_export_pixel)
+    bpy.utils.register_class(SPIO_OT_export_image)
 
 
 def unregister():
     bpy.utils.unregister_class(SPIO_OT_import_image)
+    bpy.utils.unregister_class(SPIO_OT_export_pixel)
+    bpy.utils.unregister_class(SPIO_OT_export_image)
